Refactor: Payment 도메인 분리
