+++ server/src/main/java/site/bannabe/server/domain/payments/entity/PaymentMethod.java
@@ -1,4 +1,4 @@
-package site.bannabe.server.domain.rentals.entity;
+package site.bannabe.server.domain.payments.entity;
 
 import jakarta.persistence.Converter;
 import lombok.Getter;
