+++ server/src/main/java/site/bannabe/server/domain/payments/entity/PaymentStatus.java
@@ -1,4 +1,4 @@
-package site.bannabe.server.domain.rentals.entity;
+package site.bannabe.server.domain.payments.entity;
 
 import jakarta.persistence.Converter;
 import lombok.Getter;
@@ -7,19 +7,18 @@ import site.bannabe.server.global.converter.AbstractEnumConverter;
 
 @Getter
 @RequiredArgsConstructor
-public enum PaymentType {
+public enum PaymentStatus {
 
-  RENT("대여 결제"),
-  OVERDUE("연체 결제"),
-  EXTENSION("연장 결제");
+  SUCCESS("성공"),
+  CANCEL("취소");
 
   private final String description;
 
   @Converter(autoApply = true)
-  static class EnumConverter extends AbstractEnumConverter<PaymentType> {
+  static class EnumConverter extends AbstractEnumConverter<PaymentStatus> {
 
     public EnumConverter() {
-      super(PaymentType.class);
+      super(PaymentStatus.class);
     }
 
   }
