+++ server/src/main/java/site/bannabe/server/global/utils/RandomCodeGenerator.java
@@ -3,9 +3,9 @@ package site.bannabe.server.global.utils;
 import java.security.SecureRandom;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
+import site.bannabe.server.domain.payments.entity.RentalPayments;
 import site.bannabe.server.domain.rentals.entity.RentalHistory;
 import site.bannabe.server.domain.rentals.entity.RentalItems;
-import site.bannabe.server.domain.rentals.entity.RentalPayments;
 import site.bannabe.server.domain.users.entity.Users;
 import site.bannabe.server.global.exceptions.BannabeServiceException;
 import site.bannabe.server.global.exceptions.ErrorCode;
